// src/types/flag.ts
type Flag = {
	name: string;
	image: string;
	landmass: string;
	zone: string;
	area: string;
	population: string;
	language: string;
	religion: string;
	bars: string;
	stripes: string;
	colours: string;
	red: string;
	green: string;
	blue: string;
	gold: string;
	white: string;
	black: string;
	orange: string;
	mainhue: string;
	circles: string;
	crosses: string;
	saltires: string;
	quarters: string;
	sunstars: string;
	crescent: string;
	triangle: string;
	icon: string;
	animate: string;
	text: string;
	topleft: string;
	botright: string;
};